IPC Tag Rework

+++ code/modules/mob/living/carbon/human/species/station/ipc/ipc.dm
@@ -66,7 +66,8 @@
 
 	inherent_verbs = list(
 		/mob/living/carbon/human/proc/self_diagnostics,
-		/mob/living/carbon/human/proc/change_monitor
+		/mob/living/carbon/human/proc/change_monitor,
+		/mob/living/carbon/human/proc/check_tag
 	)
 
 	flags = IS_IPC
@@ -146,28 +147,22 @@ datum/species/machine/handle_post_spawn(var/mob/living/carbon/human/H)
 	return sanitizeName(new_name, allow_numbers = 1)
 
 /datum/species/machine/proc/check_tag(var/mob/living/carbon/human/new_machine, var/client/player)
-	if (!new_machine || !player)
+	if(!new_machine || !player)
+		var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
+		if(tag)
+			tag.serial_number = uppertext(dd_limittext(md5(new_machine.real_name), 12))
+			tag.ownership_info = IPC_OWNERSHIP_COMPANY
 		return
 
-	if (establish_db_connection(dbcon))
+	var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
 
-		var/obj/item/organ/internal/ipc_tag/tag = new_machine.internal_organs_by_name[BP_IPCTAG]
-
-		var/status = TRUE
-		var/list/query_details = list("ckey" = player.ckey, "character_name" = player.prefs.real_name)
-		var/DBQuery/query = dbcon.NewQuery("SELECT tag_status FROM ss13_ipc_tracking WHERE player_ckey = :ckey: AND character_name = :character_name:")
-		query.Execute(query_details)
-
-		if (query.NextRow())
-			status = text2num(query.item[1])
-		else
-			var/DBQuery/log_query = dbcon.NewQuery("INSERT INTO ss13_ipc_tracking (player_ckey, character_name, tag_status) VALUES (:ckey:, :character_name:, 1)")
-			log_query.Execute(query_details)
-
-		if (!status)
-			new_machine.internal_organs_by_name -= BP_IPCTAG
-			new_machine.internal_organs -= tag
-			qdel(tag)
+	if(player.prefs.machine_tag_status)
+		tag.serial_number = player.prefs.machine_serial_number
+		tag.ownership_info = player.prefs.machine_ownership_status
+	else
+		new_machine.internal_organs_by_name -= BP_IPCTAG
+		new_machine.internal_organs -= tag
+		qdel(tag)
 
 /datum/species/machine/proc/update_tag(var/mob/living/carbon/human/target, var/client/player)
 	if (!target || !player)
